small refactors

+++ ToyRenderer/src/main.cpp
@@ -1,5 +1,7 @@
 #include <spdlog/spdlog.h>
 
+#include <cstddef>
+
 #include "app.h"
 #include "options.h"
 #include "registry.h"
@@ -10,7 +12,7 @@
 auto main(int argc, const char *argv[]) -> int {
   tr::Registry::load();
 
-  const tr::Options args = tr::Options::from_args(std::span(argv, argc));
+  const tr::Options args = tr::Options::from_args(std::span(argv, static_cast<std::size_t>(argc)));
   spdlog::set_level(args.debug.level);
 
   tr::App{args}.run();
